+++ js/file/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## 0.16.0
 
+### Dependency updates
+
+- @gradio/client@2.5.1
+
+## 0.16.0
+
+### Dependency updates
+
+- @gradio/client@2.5.0
+
+## 0.16.0
+
 ### Features
 
 - [#13329](https://github.com/gradio-app/gradio/pull/13329) [`7ac583a`](https://github.com/gradio-app/gradio/commit/7ac583a38a4c995d172033ff6e3700390201ff21) - Make builds go zoom zoom.  Thanks @pngwn!